Conflicts:
	Again-again/src/Control/GameControl.java
	Again-again/src/Model/Player.java
	Again-again/test/Control/FightControlTest.java

+++ Again-again/src/byui/cit260/againagain/view/TrapMenuView.java
@@ -26,8 +26,8 @@ public class TrapMenuView extends View {
     }
 
     public boolean doAction(String[] inputs) {
-        double health = 80.0;
-        health = FightControl.trapControl(80.0, 6.0, 10.0);
+        int health = 80;
+        health = FightControl.trapControl(80, 6, 10);
 
         if (health <= 0) {
             System.out.println("You're dead!");
